arm: baseline loader version support added
Soma arm cpus contains a baseline boot loader, in order to initialize the
system (x-loader for omap3, ubl for DaVinci, ...).
It is important to know its version, but the serial console is not always
available.
This patch get this information from a configurable address of the static
ram and create the baselineloader_version environment variable.

Signed-off-by: Raffaele Recalcati <raffaele.recalcati@bticino.it>